server/requirements.txt
dicttoxml==1.7.16
fastavro==1.7.3
msgpack==1.0.5
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
six==1.16.0
xmltodict==0.13.0
